isart/requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-auditlog==2.2.1
django-cors-headers==3.13.0
django-filter==22.1
django-import-export==3.0.1
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
greenlet==2.0.1
jsonfield==3.1.0
MarkupPy==1.14
mysqlclient==2.1.1
numpy==1.23.5
odfpy==1.4.1
openpyxl==3.0.10
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
tablib==3.2.1
xlrd==2.0.1
xlwt==1.3.0
